+++ ResourceKit/Sources/Font.swift
@@ -6,8 +6,7 @@
 //  Copyright © 2023 chansoo.MOIT. All rights reserved.
 //
 
-// TODO: extension 메서드 위치 이동 후 Foundation으로 변경 필요
-import UIKit
+import Foundation
 
 extension ResourceKitFontFamily {
     public static let h1 = Pretendard.bold.font(size: 36)
@@ -32,34 +31,3 @@ extension ResourceKitFontFamily {
 		else { return 0 }
 	}
 }
-
-// TODO: 추후 Utils로 위치 이동 필요
-public extension UILabel {
-	func setTextWithParagraphStyle(
-		text: String,
-		alignment: NSTextAlignment = .left,
-		font: ResourceKitFontConvertible.Font,
-		textColor: UIColor
-	) {
-		let paragraphStyle = NSMutableParagraphStyle()
-		paragraphStyle.alignment = alignment
-		
-		let fontHeight = ResourceKitFontFamily.lineHeight(of: font)
-		paragraphStyle.maximumLineHeight = fontHeight
-		paragraphStyle.minimumLineHeight = fontHeight
-		
-		let attributes: [NSAttributedString.Key : Any] = [
-			.paragraphStyle : paragraphStyle,
-			.font: font,
-			.foregroundColor: textColor,
-			.baselineOffset: (fontHeight - font.lineHeight) / 4
-		]
-		
-		debugPrint(font.lineHeight)
-		
-		let attrString = NSAttributedString(string: text,
-											attributes: attributes)
-		self.attributedText = attrString
-	}
-}
-
